move the measurement class into the measurement models header
divorce measurements from filters
